services/shipment-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
boto3==1.28.62
kafka-python==2.0.2
aiokafka==0.8.1
ecdsa==0.18.0
psycopg2-binary
asyncpg
sqlalchemy==2.0.23
PyJWT==2.8.0
python-multipart==0.0.6
